Upstream changed a lot; simplify clock setup code
Structure of lookups for clock ratios etc. has changed significantly in
the 4.7.2 driver, to the point that the original 4.4 patch doesn't come
close to applying cleanly.

But, it appears that to address the underlying issue here (the MCLK not
getting set up) we can simply intercept the driver's
es8328_hw_params function (when it realizes that the sysclk_constraints
pointer is null), and force the clock value there.

This _seems_ to work. I may have misunderstood the logic though, so this
should be checked.