Fix eDP adaptor boot problem
In 4.7.2, the call in imx_ldb_bind to find the output port, followed
by looking up the endpoint of that port, was replaced by a call to
of_graph_get_endpoint_by_regs, to get the endpoint in a single step.

Unfortunately, the new call causes the kernel to die during the boot
process. Revert to original two-stage process for now.

Thanks to veleiro for reporting.